Adding Anarctic/Arctic map MVP
This is not intended for long-term use but to verify we can configure a vanilla OpenLayers map with Antarctic/Arctic data before integrating this into Vue/VueLayers.

This commit adds basic maps using WMS (not WMTS).

As this is a temporary service, I am not intending to add this as a Nomad task, and this will therefore only work locally.