Adding projection control
Note, though this control works in that it will change the map projection, it won’t succeeed due to an edge case in VueLayers (related to the async way the map updates).

This is being addressed but until then only the initial projection is supported. Changing the projection will cause a runtime error and crash the app.