Fixing base layers on projection change
Previously when the projection was changed an appropriate base layer was added to `active_layers` and was reflected in the Vue dev tools but didn’t get drawn on the map.

Changing this so the layers aren’t changed until the map has updated has resolved this problem, so I think this was because the layers were changed before the map had finished reprojecting.