Adding the ability to remove granules from the map
This required adding the layer and granule IDs to each active layer, so that we can lookup whether a given granule is being displayed when we try to remove it.

Currently we use both the granule and layer ID to do this but in reality we just use the granule ID since granules are unique to a layer. I’ve chosen to be stricter until the way we handle granules and layers on the map is better established.