Adding deployment package for frontend application
Uses GitLab CD to publish the bundle produced by Parcel as a zipped generic package in the GitLab Package Registry (which is per-project).

Two packages are added, the default output from Parcel, and another without source maps, to reduce the package size (but hampering debugging). This second package is intended for the situation that new releases need to be downloaded onto the ship and therefore bandwidth is at a premium.