Initial work on new components and passing events
This splits up the previously monolithic app module into a set of smaller modules for product selection, map controls and various metadata panels.

Events are used to update other components as needed (for example when the CRS is changed in the map controls component, the hemisphere filter is changed in the product switcher component).

This is currently not done in the best way (all events passed to parent and then passed back down to relevant children, rather than being passed between siblings).

This commit only adds enough controls, elements and events to prove this concept works. They will be fleshed out later.