Adding support for real products
Initially with minimal properties to prove we can pass objects around.

Currently the initially active products list is hard-coded with the base layer for the North polar projection. This is ok as that’s the default map projection but we need a method to determine a new list of active layers when the projection changes.