Linking colour scheme to map
Specifically, a computed style suffix (`-day` or `-night`) is added to all style names based on the application colour scheme (light, dark or system).

This got a bit tricky with the ‘system’ mode as need to resolve the current state without becoming fixed on the resolved value.

This implementation is not as clean as I’d like (it seems overly complex) but it should work.