Removing legacy UI
This essentially replaces the older ‘app’ components with the ‘app2’ componments leaving a single set of components.

The new UI developed in the app2 components has now significantly overtaken and evolved the legacy UI - to the point the legacy UI is no longer relevant or useful.

Removing these old components now, simplifies the application and avoids technical debt.