Adding footprint to granule selection support
This converts selected footprints into selected granules within a product component. This completes the minimum needed to select footprints on the map, pass them to the product and pass back granules to be shown on the map.

Note: There is a bug currently where the selected footprint appears deselected when the map layers are redrawn. I’ve tried a few options for addressing this crudely but I think I will need to review how the footprints vector layer is added to fix that.