Fixing map layers not being updated when values change
Previously we reset the layers array whenever the active products/granules list changed - this had the implict effect of re-rendering the component, and therefore update the map when properties such as opacity changed.

When we stopped doing this to preserve footprint layers (to preserve selected features), these map updates stopped. The properties within the component were updated, but the component wasn’t re-rendered.

For now I’ve fixed this by making this re-render explict at the end of the layers init method.

However this isn’t best practice (we should let vue determine when it needs to (re-)render the component so needs a longer term in the future.