Fixing bug when changing filter selection
Previously we assumed when changing time/date filter range that we’d always get at least one granule, and so could select the last item in the granules array.

This is not always the case (i.e. selecting a period in which there are no granules). Therefore this logic needs to be conditional and where no granules are present then no granules should be selected.